vac_backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.42.2
cryptography==44.0.2
Django==5.1.7
django-ckeditor==6.7.2
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.9
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.0
python-http-client==3.3.7
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
